Remove model.py and refactor its dependents
