+++ tests/sources/test_base.py
@@ -1,4 +1,3 @@
-import datetime
 from pathlib import Path
 from typing import Iterable
